maktab

+++ order_items/delete.php
@@ -14,13 +14,19 @@ $data = $conn->prepare($sql);
 $data->execute([$id]);
 $order_item = $data->fetch();
 
-$sql = "DELETE FROM order_item WHERE id = ?";
-$data = $conn->prepare($sql);
-$data->execute([$id]);
+if($order_item){
 
-$data = $conn->prepare("UPDATE books SET quantity = quantity + 1  WHERE id = ?
-");
-$data->execute([$id]);
+    $data = $conn->prepare("
+        UPDATE books 
+        SET quantity = quantity + 1 
+        WHERE id = ?
+    ");
+    $data->execute([$order_item['book_id']]);
+
+    $sql = "DELETE FROM order_item WHERE id = ?";
+    $data = $conn->prepare($sql);
+    $data->execute([$id]);
+}
 
 header("Location: ../orders/show.php?id=" . $order_item['order_id']);
 exit();
